extract long-running tests to separate file, update javadocs
