+++ src/test/java/io/iron/ironmq/IronMQTest.java
@@ -165,24 +165,6 @@ public class IronMQTest {
     }
 
     /**
-     * Test shows how to increase time of message reservation
-     * Expected that:
-     * - message will be available after 5 seconds (initial timeout is 5 seconds)
-     * @throws IOException
-     * @throws InterruptedException
-     */
-    @Test
-    public void testReserveMessageWithTimeout() throws IOException, InterruptedException {
-        queue.push("Test message");
-        Message reservedFirstTime = queue.reserve(1, 30).getMessage(0);
-
-        Thread.sleep(32000);
-        Messages reservedAgain = queue.reserve(1);
-        Assert.assertEquals(1, reservedAgain.getSize());
-        Assert.assertEquals(reservedAgain.getMessage(0).getId(), reservedFirstTime.getId());
-    }
-
-    /**
      * This test shows how to peek a message from a queue
      * Expected that
      * - Messsage has id and body
@@ -333,65 +315,6 @@ public class IronMQTest {
 
         messages.getMessage(0).setReservationId(null);
         queue.deleteMessages(messages);
-    }
-
-    /**
-     * Test shows how to increase time of message reservation
-     * Expected that:
-     * - message will be available after 5 seconds (initial timeout is 5 seconds)
-     * @throws IOException
-     * @throws InterruptedException
-     */
-    @Test
-    public void testTouchMessage() throws IOException, InterruptedException {
-        queue.push("Test message");
-        Message message = queue.reserve(1, 30).getMessage(0);
-
-        Thread.sleep(25000);
-        queue.touchMessage(message);
-        Thread.sleep(10000);
-        Assert.assertEquals(0, queue.reserve(1).getSize());
-    }
-
-    /**
-     * Test shows how to increase time of message reservation
-     * Expected that:
-     * - message will be available after 5 seconds (initial timeout is 5 seconds)
-     * @throws IOException
-     * @throws InterruptedException
-     */
-    @Test
-    public void testTouchMessageWithTimeout() throws IOException, InterruptedException {
-        queue.push("Test message");
-        Message message = queue.reserve(1, 30).getMessage(0);
-        queue.touchMessage(message, 40);
-
-        Thread.sleep(35000);
-        Assert.assertEquals(0, queue.reserve(1).getSize());
-        Thread.sleep(10000);
-        Assert.assertEquals(1, queue.reserve(1).getSize());
-    }
-
-    /**
-     * Test shows how to touch a message multiple times.
-     * Expected that:
-     * - after second touch call message will have new reservation id
-     * - new reservation id will not equal to old one
-     * @throws IOException
-     * @throws InterruptedException
-     */
-    @Test
-    public void testTouchMessageTwice() throws IOException, InterruptedException {
-        queue.push("Test message");
-        Message message = queue.reserve(1, 5).getMessage(0);
-
-        Thread.sleep(3500);
-        MessageOptions options1 = queue.touchMessage(message);
-        Thread.sleep(3500);
-        MessageOptions options2 = queue.touchMessage(message);
-
-        Assert.assertFalse(options1.getReservationId().equals(options2.getReservationId()));
-        Assert.assertEquals(message.getReservationId(), options2.getReservationId());
     }
 
     /**
